refactor: extract LocaleOption
