+++ src/components/LocaleSelect/LocaleSelect.scss
@@ -11,7 +11,7 @@
     font-weight: 700;
     color: #444;
     border: 1px solid #aaa;
-    border-radius: .5em;
+    border-radius: 0.5em;
 
     &:hover,
     &:focus {
@@ -23,5 +23,4 @@
   &--item {
     background-color: #fff;
   }
-
 }
