+++ src/pages/MainPage.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const MainPage = () => <div>main</div>;
+
+export default MainPage;
